Feature/productivity service tests

+++ productivity-service/src/main/java/com/ilyap/productivityservice/controller/ProductivityController.java
@@ -2,10 +2,12 @@ package com.ilyap.productivityservice.controller;
 
 import com.ilyap.productivityservice.model.dto.ProductivityCreateUpdateDto;
 import com.ilyap.productivityservice.model.dto.ProductivityReadDto;
+import com.ilyap.productivityservice.security.ProductivityPermissionHandler;
 import com.ilyap.productivityservice.service.ProductivityService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.oauth2.server.resource.authentication.JwtAuthenticationToken;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -14,34 +16,57 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
 import reactor.core.publisher.Mono;
 
 import java.util.UUID;
 
 @RestController
-@RequestMapping("/api/v1/productivity/{productivityId}")
+@RequestMapping("/api/v1/productivity")
 @RequiredArgsConstructor
 public class ProductivityController {
 
     private final ProductivityService productivityService;
+    private final ProductivityPermissionHandler productivityPermissionHandler;
 
-    @PreAuthorize("@productivityPermissionHandler.isProductivityUser(#productivityId, principal.username)")
-    @GetMapping
-    public Mono<ProductivityReadDto> findById(@PathVariable UUID productivityId) {
-        return productivityService.findById(productivityId);
+    @GetMapping("/{productivityId}")
+    public Mono<ProductivityReadDto> findById(@PathVariable UUID productivityId,
+                                              JwtAuthenticationToken authentication) {
+        return productivityPermissionHandler.isProductivityUser(productivityId, authentication.getName())
+                .flatMap(isAuthorized -> {
+                    if (isAuthorized) {
+                        return productivityService.findById(productivityId);
+                    } else {
+                        return Mono.error(new ResponseStatusException(HttpStatus.FORBIDDEN, "User is not authorized"));
+                    }
+                });
     }
 
-    @PreAuthorize("@productivityPermissionHandler.isProductivityUser(#productivityId, principal.username)")
-    @PutMapping
+    @PutMapping("/{productivityId}")
     public Mono<ProductivityReadDto> update(@PathVariable UUID productivityId,
-                                            @Validated @RequestBody ProductivityCreateUpdateDto createUpdateDto) {
-        return productivityService.update(productivityId, createUpdateDto);
+                                            @Validated @RequestBody ProductivityCreateUpdateDto createUpdateDto,
+                                            JwtAuthenticationToken authentication) {
+        return productivityPermissionHandler.isProductivityUser(productivityId, authentication.getName())
+                .flatMap(isAuthorized -> {
+                    if (isAuthorized) {
+                        return productivityService.update(productivityId, createUpdateDto);
+                    } else {
+                        return Mono.error(new ResponseStatusException(HttpStatus.FORBIDDEN, "User is not authorized"));
+                    }
+                });
     }
 
-    @PreAuthorize("@productivityPermissionHandler.isProductivityUser(#productivityId, principal.username)")
-    @DeleteMapping
-    public Mono<ResponseEntity<Void>> delete(@PathVariable UUID productivityId) {
-        return productivityService.delete(productivityId)
-                .then(Mono.just(ResponseEntity.noContent().build()));
+    @DeleteMapping("/{productivityId}")
+    public Mono<ResponseEntity<Void>> delete(@PathVariable UUID productivityId,
+                                             JwtAuthenticationToken authentication) {
+        return productivityPermissionHandler.isProductivityUser(productivityId, authentication.getName())
+                .flatMap(isAuthorized -> {
+                    if (isAuthorized) {
+                        return productivityService.delete(productivityId)
+                                .then(Mono.just(ResponseEntity.noContent().build()));
+                    } else {
+                        return Mono.error(new ResponseStatusException(HttpStatus.FORBIDDEN, "User is not authorized"));
+                    }
+                });
     }
 }
